move tracer config to its own file
