+++ source/extensions/tracers/dynamic_ot/BUILD
@@ -30,7 +30,7 @@ envoy_cc_library(
     hdrs = ["config.h"],
     deps = [
         ":dynamic_opentracing_driver_lib",
+        "//include/envoy/server:tracer_config_interface",
         "//source/extensions/tracers:well_known_names",
-        "//source/server:configuration_lib",
     ],
 )
