+++ source/extensions/tracers/lightstep/BUILD
@@ -20,10 +20,10 @@ envoy_cc_library(
     ],
     external_deps = ["lightstep"],
     deps = [
+        "//source/common/config:utility_lib",
         "//source/common/tracing:http_tracer_lib",
         "//source/extensions/tracers:well_known_names",
         "//source/extensions/tracers/common/ot:opentracing_driver_lib",
-        "//source/server:configuration_lib",
     ],
 )
 
@@ -33,7 +33,7 @@ envoy_cc_library(
     hdrs = ["config.h"],
     deps = [
         ":lightstep_tracer_lib",
+        "//include/envoy/server:tracer_config_interface",
         "//source/extensions/tracers:well_known_names",
-        "//source/server:configuration_lib",
     ],
 )
