+++ source/extensions/tracers/lightstep/config.h
@@ -3,8 +3,7 @@
 #include <string>
 
 #include "envoy/server/instance.h"
-
-#include "server/configuration_impl.h"
+#include "envoy/server/tracer_config.h"
 
 namespace Envoy {
 namespace Extensions {
