+++ source/extensions/tracers/zipkin/BUILD
@@ -48,6 +48,7 @@ envoy_cc_library(
         "//source/common/common:enum_to_int",
         "//source/common/common:hex_lib",
         "//source/common/common:utility_lib",
+        "//source/common/config:utility_lib",
         "//source/common/http:header_map_lib",
         "//source/common/http:message_lib",
         "//source/common/http:utility_lib",
@@ -56,7 +57,6 @@ envoy_cc_library(
         "//source/common/singleton:const_singleton",
         "//source/common/tracing:http_tracer_lib",
         "//source/extensions/tracers:well_known_names",
-        "//source/server:configuration_lib",
     ],
 )
 
@@ -65,8 +65,8 @@ envoy_cc_library(
     srcs = ["config.cc"],
     hdrs = ["config.h"],
     deps = [
+        ":zipkin_lib",
+        "//include/envoy/server:tracer_config_interface",
         "//source/extensions/tracers:well_known_names",
-        "//source/extensions/tracers/zipkin:zipkin_lib",
-        "//source/server:configuration_lib",
     ],
 )
