+++ source/server/configuration_impl.h
@@ -10,13 +10,11 @@
 #include <utility>
 
 #include "envoy/config/bootstrap/v2/bootstrap.pb.h"
-#include "envoy/config/trace/v2/trace.pb.h"
 #include "envoy/http/filter.h"
 #include "envoy/network/filter.h"
 #include "envoy/server/configuration.h"
 #include "envoy/server/filter_config.h"
 #include "envoy/server/instance.h"
-#include "envoy/tracing/http_tracer.h"
 
 #include "common/common/logger.h"
 #include "common/json/json_loader.h"
@@ -26,39 +24,6 @@
 namespace Envoy {
 namespace Server {
 namespace Configuration {
-
-/**
- * Implemented by each Tracer and registered via Registry::registerFactory() or the convenience
- * class RegisterFactory.
- */
-class TracerFactory {
-public:
-  virtual ~TracerFactory() {}
-
-  /**
-   * Create a particular HttpTracer implementation. If the implementation is unable to produce an
-   * HttpTracer with the provided parameters, it should throw an EnvoyException in the case of
-   * general error or a Json::Exception if the json configuration is erroneous. The returned
-   * pointer should always be valid.
-   * @param json_config supplies the general json configuration for the HttpTracer
-   * @param server supplies the server instance
-   */
-  virtual Tracing::HttpTracerPtr
-  createHttpTracer(const envoy::config::trace::v2::Tracing& configuration, Instance& server) PURE;
-
-  /**
-   * @return ProtobufTypes::MessagePtr create empty config proto message for v2. The tracing
-   *         config, which arrives in an opaque google.protobuf.Struct message, will be converted to
-   *         JSON and then parsed into this empty proto.
-   */
-  virtual ProtobufTypes::MessagePtr createEmptyConfigProto() PURE;
-
-  /**
-   * Returns the identifying name for a particular implementation of tracer produced by the
-   * factory.
-   */
-  virtual std::string name() PURE;
-};
 
 /**
  * Implemented for each Stats::Sink and registered via Registry::registerFactory() or
